streaming-iot-pipeline/requirements.txt
annotated-types==0.7.0
apache-beam==2.57.0
async-timeout==4.0.3
attrs==23.2.0
avro==1.11.3
cachetools==4.2.4
certifi==2024.7.4
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
Faker==26.0.0
fastavro==1.9.5
fasteners==0.19
future==0.16.0
gapic-google-cloud-pubsub-v1==0.15.4
google-api-core==2.19.1
google-apitools==0.5.31
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.59.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.25.0
google-cloud-bigtable==2.24.0
google-cloud-core==2.4.1
google-cloud-dataflow==2.2.0
google-cloud-dataflow-client==0.8.11
google-cloud-datastore==2.19.0
google-cloud-dlp==3.18.1
google-cloud-language==2.13.4
google-cloud-pubsub==2.22.0
google-cloud-pubsublite==1.11.0
google-cloud-recommendations-ai==0.10.11
google-cloud-resource-manager==1.12.4
google-cloud-spanner==3.47.0
google-cloud-storage==2.17.0
google-cloud-videointelligence==2.13.4
google-cloud-vision==3.7.3
google-crc32c==1.5.0
google-gax==0.15.16
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
googledatastore==7.0.1
grpc-google-iam-v1==0.13.1
grpc-interceptor==0.15.4
grpcio==1.64.1
grpcio-status==1.62.2
hdfs==2.7.3
httplib2==0.22.0
idna==3.7
Js2Py==0.74
jsonpickle==3.2.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
load-dotenv==0.1.0
mock==2.0.0
numpy==1.26.4
oauth2client==3.0.0
objsize==0.7.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pbr==6.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-pubsub-v1==0.15.4
proto-plus==1.24.0
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydot==1.4.2
pyjsparser==2.7.1
pymongo==4.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==3.13
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
rpds-py==0.19.0
rsa==4.9
shapely==2.0.5
six==1.16.0
sqlparse==0.5.0
typing==3.6.6
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
zstandard==0.22.0
